refactor [data.json 1/2] Extract data.json computation from the vue component
So we can use it on the script
